refactor: add angular v19 and v20 adapters$ $ Refs: #7112
