+++ package.json
@@ -5,7 +5,7 @@
 		"husky": "9.1.7",
 		"lerna": "8.2.2",
 		"license-report": "6.8.0",
-		"lint-staged": "16.1.0",
+               "lint-staged": "16.1.2",
 		"npm-check-updates": "18.0.1",
 		"npm-run-all2": "8.0.4",
 		"prettier": "3.5.3"
