--- packages/adapters/angular/v15/README.md
@@ -1,19 +1,0 @@
-# Angular-Adapter
-
-[![npm](https://img.shields.io/npm/v/@public-ui/angular-v15)](https://www.npmjs.com/package/@public-ui/components)
-[![license](https://img.shields.io/npm/l/@public-ui/angular-v15)](https://github.com/public-ui/kolibri/blob/main/LICENSE)
-[![downloads](https://img.shields.io/npm/dt/@public-ui/angular-v15)](https://www.npmjs.com/package/@public-ui/angular-v15)
-[![issues](https://img.shields.io/github/issues/public-ui/kolibri)](https://github.com/public-ui/kolibri/issues)
-[![pull requests](https://img.shields.io/github/issues-pr/public-ui/kolibri)](https://github.com/public-ui/kolibri/pulls)
-[![size](https://img.shields.io/bundlephobia/min/@public-ui/angular-v15)](https://bundlephobia.com/result?p=@public-ui/angular-v15)
-![contributors](https://img.shields.io/github/contributors/public-ui/kolibri)
-
-Das [**Angular**](https://angular.io)-Modul ist der Framework-Adapter für die Komponenten-Bibliothek.
-
-Mehr zur **Modularisierung** kann im [Architekturkonzept](https://public-ui.github.io/docs/concepts/architecture) nachgelesen werden.
-
-Mehr zum **Projekt** kann in der [README](https://public-ui.github.io/docs) nachgelesen werden.
-
-## Referenzen
-
-- <https://entwickler.de/angular/der-teufel-steckt-im-detail-001/>
