--- packages/adapters/angular/v15/tsconfig.json
@@ -1,38 +1,0 @@
-{
-	"angularCompilerOptions": {
-		"annotateForClosureCompiler": false,
-		"strictMetadataEmit": true,
-		"flatModuleOutFile": "core.js",
-		"flatModuleId": "@public-ui/angular",
-		"skipTemplateCodegen": true,
-		"fullTemplateTypeCheck": false,
-		"enableIvy": true
-	},
-	"compilerOptions": {
-		"alwaysStrict": true,
-		"strict": true,
-		"allowSyntheticDefaultImports": true,
-		"allowUnreachableCode": false,
-		"declaration": true,
-		"declarationDir": "dist",
-		"experimentalDecorators": true,
-		"forceConsistentCasingInFileNames": true,
-		"lib": ["dom", "es2017"],
-		"module": "ESNext",
-		"moduleResolution": "node",
-		"noImplicitAny": true,
-		"noImplicitReturns": true,
-		"noUnusedLocals": false,
-		"noUnusedParameters": true,
-		"outDir": "dist",
-		"pretty": true,
-		"removeComments": false,
-		"importHelpers": true,
-		"rootDir": "src",
-		"strictPropertyInitialization": false,
-		"target": "es2015",
-		"skipLibCheck": true
-	},
-	"files": ["src/index.ts"],
-	"exclude": ["node_modules"]
-}
