--- packages/adapters/angular/v16/package.json
@@ -1,89 +1,0 @@
-{
-	"name": "@public-ui/angular-v16",
-	"version": "2.2.11",
-	"license": "EUPL-1.2",
-	"homepage": "https://public-ui.github.io",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/public-ui/kolibri"
-	},
-	"bugs": {
-		"url": "https://github.com/public-ui/kolibri/issues",
-		"email": "kolibri@itzbund.de"
-	},
-	"author": {
-		"name": "Informationstechnikzentrum Bund",
-		"email": "kolibri@itzbund.de"
-	},
-	"sideEffects": false,
-	"description": "Angular (v16) framework adapter for KoliBri - The accessible HTML-Standard.",
-	"keywords": [
-		"accessibility",
-		"accessible",
-		"bitv",
-		"framework",
-		"library",
-		"designsystem",
-		"design",
-		"system",
-		"web components",
-		"webcomponents",
-		"aria",
-		"wai",
-		"axe",
-		"custom elements",
-		"styleguide",
-		"style",
-		"guide",
-		"ui",
-		"html",
-		"css",
-		"web",
-		"a11y",
-		"w3c",
-		"webstandard",
-		"wcag",
-		"angular"
-	],
-	"scripts": {
-		"clean": "lerna exec --stream -- git clean -f -d -X && git clean -f -d -X",
-		"build:cjs": "tsc -m commonjs --outDir cjs",
-		"build:clean": "rimraf cjs esm types umd",
-		"build:esm": "tsc -m esnext --outDir esm",
-		"build:ngc": "node ng-module.js && ngc -p .",
-		"build:tsc": "tsc -p .",
-		"build:types": "tsc -d --outDir types && rimraf types/*.js types/*.map types/**/*.js types/**/*.map",
-		"build:umd": "tsc -m umd --outDir umd",
-		"build": "npm run build:clean && npm run build:ngc"
-	},
-	"main": "./dist/index.js",
-	"module": "./dist/index.js",
-	"types": "./dist/index.d.ts",
-	"exports": {
-		"types": "./dist/index.d.ts",
-		"require": "./dist/index.js",
-		"import": "./dist/index.js"
-	},
-	"devDependencies": {
-		"@angular/common": "16.2.12",
-		"@angular/compiler": "16.2.12",
-		"@angular/compiler-cli": "16.2.12",
-		"@angular/core": "16.2.12",
-		"@public-ui/components": "workspace:*",
-		"@types/minimatch": "5.1.2",
-		"@types/minimist": "1.2.5",
-		"@types/normalize-package-data": "2.4.4",
-		"prettier": "3.5.3",
-		"rimraf": "6.0.1",
-		"rxjs": "7.8.2",
-		"typescript": "5.1.6",
-		"zone.js": "0.13.3"
-	},
-	"peerDependencies": {
-		"@angular/core": "16.2.12",
-		"@public-ui/components": "workspace:*"
-	},
-	"files": [
-		"dist"
-	]
-}
