+++ packages/adapters/angular/v19/.gitignore
