+++ packages/adapters/angular/v19/.npmignore
