+++ packages/adapters/angular/v19/LICENSE
