+++ packages/adapters/angular/v20/.gitignore
@@ -1,0 +1,1 @@
+/src/
