+++ packages/adapters/angular/v20/.npmignore
@@ -1,0 +1,1 @@
+/src/
