+++ packages/adapters/angular/v20/LICENSE
