+++ packages/samples/angular/package.json
@@ -10,24 +10,24 @@
 		"watch": "ng build --watch --configuration development"
 	},
 	"dependencies": {
-		"@angular/animations": "18.2.13",
-		"@angular/common": "18.2.13",
-		"@angular/compiler": "18.2.13",
-		"@angular/core": "18.2.13",
-		"@angular/forms": "18.2.13",
-		"@angular/platform-browser": "18.2.13",
-		"@angular/platform-browser-dynamic": "18.2.13",
-		"@angular/router": "18.2.13",
-		"@public-ui/angular-v18": "workspace:*",
+               "@angular/animations": "20.0.3",
+               "@angular/common": "20.0.3",
+               "@angular/compiler": "20.0.3",
+               "@angular/core": "20.0.3",
+               "@angular/forms": "20.0.3",
+               "@angular/platform-browser": "20.0.3",
+               "@angular/platform-browser-dynamic": "20.0.3",
+               "@angular/router": "20.0.3",
+               "@public-ui/angular-v20": "workspace:*",
 		"@public-ui/components": "workspace:*",
 		"@public-ui/theme-default": "workspace:*",
 		"rxjs": "7.8.2",
 		"zone.js": "0.14.10"
 	},
 	"devDependencies": {
-		"@angular-devkit/build-angular": "18.2.20",
-		"@angular/cli": "18.2.20",
-		"@angular/compiler-cli": "18.2.13",
+               "@angular-devkit/build-angular": "20.0.2",
+               "@angular/cli": "20.0.2",
+               "@angular/compiler-cli": "20.0.3",
 		"@types/jasmine": "5.1.8",
 		"jasmine-core": "5.8.0",
 		"karma": "6.4.4",
