+++ scripts/deprecate-add.sh
@@ -1,7 +1,7 @@
-npm deprecate @public-ui/angular-v15@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
-npm deprecate @public-ui/angular-v16@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
 npm deprecate @public-ui/angular-v17@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
 npm deprecate @public-ui/angular-v18@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
+npm deprecate @public-ui/angular-v19@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
+npm deprecate @public-ui/angular-v20@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
 npm deprecate @public-ui/hydrate@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
 npm deprecate @public-ui/kolibri-cli@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
 npm deprecate @public-ui/preact@$1 "Wir empfehlen die Verwendung einer neueren Version (>=$2)."
