+++ scripts/dist-tags.sh
@@ -1,7 +1,7 @@
-npm dist-tag $1 @public-ui/angular-v15@$2 $3
-npm dist-tag $1 @public-ui/angular-v16@$2 $3
 npm dist-tag $1 @public-ui/angular-v17@$2 $3
 npm dist-tag $1 @public-ui/angular-v18@$2 $3
+npm dist-tag $1 @public-ui/angular-v19@$2 $3
+npm dist-tag $1 @public-ui/angular-v20@$2 $3
 npm dist-tag $1 @public-ui/hydrate@$2 $3
 npm dist-tag $1 @public-ui/kolibri-cli@$2 $3
 npm dist-tag $1 @public-ui/preact@$2 $3
